fix(mod): thx goeng - price token 분리, 아이스하키 하키로 바꿔치기
