+++ src/libs/types/item.ts
@@ -6,7 +6,7 @@ export const PRODUCT_TYPES_MAP = {
   SOCCER: '축구',
   BASKETBALL: '농구',
   BASEBALL: '야구',
-  HOCKEY: '아이스하키',
+  HOCKEY: '하키',
   VARSITY_JACKET: '과잠',
   ACCESSORY: '악세사리',
   SELF_MADE: '자체제작',
